+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+- Added `Deref` and `DerefMut` impls for `Single`, `TrySingle`, `Has<Q>`.
+
 ## 0.5.0 - 2024-04-07
 
 - Redesign component access checking. Complex queries will now pass the component access checker in more cases.